moved Notification to DeRecHelperNotification (#40 fix)
